Touch Fish mostly works now, the problem is now live activity number will only change when an update is received from the app, but app whil in background will halt

+++ LiveActivityTest/LiveActivityAttributes.swift
@@ -10,7 +10,7 @@ import ActivityKit
 struct MyAttributes: ActivityAttributes {
     public struct ContentState: Codable, Hashable {
         var status: String
-        var counter: Int
+        var counter: Double
     }
 
     var name: String
